backend/requirements.txt
aiodns==2.0.0
aiofile==3.3.3
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
brotlipy==0.7.0
caio==0.6.3
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
Django==3.1.5
django-cors-headers==3.6.0
django-mongoengine==0.4.4
django-restframework==0.0.1
djangorestframework==3.12.2
dnspython==2.1.0
idna==2.10
mongoengine==0.22.1
multidict==5.1.0
passlib==1.7.4
pycares==3.1.1
pycparser==2.20
pycryptodomex==3.9.9
PyJWT==1.7.1
pymongo==3.11.2
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stream-chat==3.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
